+++ app/server/lib/sendAppPage.ts
@@ -3,10 +3,10 @@ import {isAffirmative} from 'app/common/gutil';
 import {getTagManagerSnippet} from 'app/common/tagManager';
 import {Document} from 'app/common/UserAPI';
 import {SUPPORT_EMAIL} from 'app/gen-server/lib/HomeDBManager';
-import {appSettings} from 'app/server/lib/AppSettings';
 import {isAnonymousUser, isSingleUserMode, RequestWithLogin} from 'app/server/lib/Authorizer';
 import {RequestWithOrg} from 'app/server/lib/extractOrg';
 import {GristServer} from 'app/server/lib/GristServer';
+import {getTemplateOrg} from 'app/server/lib/gristSettings';
 import {getSupportedEngineChoices} from 'app/server/lib/serverUtils';
 import {readLoadedLngs, readLoadedNamespaces} from 'app/server/localization';
 import * as express from 'express';
@@ -152,18 +152,6 @@ export function makeSendAppPage(opts: {
       );
     resp.status(options.status).type('html').send(content);
   };
-}
-
-export function getTemplateOrg() {
-  let org = appSettings.section('templates').flag('org').readString({
-    envVar: 'GRIST_TEMPLATE_ORG',
-  });
-  if (!org) { return null; }
-
-  if (process.env.GRIST_ID_PREFIX) {
-    org += `-${process.env.GRIST_ID_PREFIX}`;
-  }
-  return org;
 }
 
 function shouldSupportAnon() {
